Increase default object file pool size and improve error
It's possible on machines that create a lot of very short lived processes to evict an object file from the pool before it was possible to be used. In these cases the best solution is to increase the object pool size since not increasing it would lead to extremely high cache misses which would also be very bad.