tests: Add integration tests for mixed-mode unwinding
aka JIT <-> DWARF unwinding. This also tests that the symbolization using
perfdump works. Note that asserting on the JIT stacks is temporarily disabled in
CI as it's consistently failing and we have other features to work on
that are higher priority. Landing this now will at least ensure that
this is tested locally. Will make sure that this CI exception is lifted
once we have more time.

Also changed the running time for the profiler during non-CI runs to
reduce feedback cycles.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>