Ensure we are running under root
This was previously discussed in
https://github.com/parca-dev/parca-agent/pull/719.

While ideally we could check for the individual capabilities this is
fraught with issues: different kernel revisions might require different
capabilities, over time we are adding new features that might require or
drop the needed capabilities, and in general, capabilities don't seem to
be a feature that many users are intimately familiar with.

Checking for the superuser is a reasonable thing to do as it will ensure
we can run, and it's how every single deployment that we know of
executes this software.

We can always re-evaluate this in the future, but in the meantime let's
ensure that running the Agent does not result in surprises
(https://github.com/parca-dev/parca-agent/issues/1979, but we've seen
this happening multiple times).

As there were concerns regarding this check being to broad, I've added a
escape hatch, which is disabled by default
(`--allow-running-as-non-root`).

Test Plan
========

**no root w/ escape hatch**
```
$ dist/parca-agent --allow-running-as-non-root
level=error name=parca-agent ts=2023-09-07T08:39:27.529651017Z caller=main.go:345 err="bpf check: perf event program type not supported: errno 0\nstack trace map type not supported: errno 0\nhash map type not supported: errno 0"
```

**no root**
```
$ dist/parca-agent
level=error name=parca-agent ts=2023-09-07T08:39:34.003842932Z caller=main.go:345 err="superuser (root) is required to run Parca Agent to load and manipulate BPF programs"
```

**root**

```
$ sudo dist/parca-agent
```

It works fine