pkg/perf: More robust perf map parsing
Perf map files have fundamental race conditions between the producer and
consumer and it's never guaranteed for lines to be complete or correct.
Therefore we need to be more rigorous about validating lines and
expecting the worst. This patch was originally triggered by a customer
seeing panics with out of bounds accesses, which have been simulated in
the added regression tests.