Reduce size of object pool
512 elf.Files (+ other metadata) can really balloon memory usage. Assuming
each entry is ~10 MB would result in a heap usage of at least 5 GB. Some
of our users reported 10 GB in heap usage in the pool which is
completely unreasonable.

In the future we want to update this to have an adaptive mechanism so we
can use the cache as much as possible while keeping memory usage at bay.

Right now having a pool to small will result in more files being opened,
(elf.Open) which results in lots of allocations and high CPU usage.
Hence we have a tradeoff of CPU + allocations vs heap usage.

Test Plan
=========

Ran overnight without issues.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>