bpf: rate limit events
We've seen some reports of boxes with overloaded CPU usage (unrelated to
the Agent) that significantly less frames that we expected. The profiles
are already incomplete before they are on the wire and after some
investigation and carefully looking at the metrics, we could not spot
anything abnormal (no normalization errors, no dropped samples, etc).

The current working theory is that the BPF events channel (we use a perf
buffer atm) might be writing faster than we are reading, so we are
losing events. While instrumenting the code to find whether this was the
case, it seemed that libbpfgo's lost events channel never receives
anything. This is something we are debugging in parallel as we really
want to fix this and have metrics on lost events.

To test the hyphotesis of a overloaded CPU producing lots of events
(lots of forks) I tested it on my machine and manually instrumented both
BPF and userspace to count events sent vs. received and the differences
were abysmal.

In this commit a small rate limiting for events is added with the aim of
reducing the chances of losing events.

Future improvements
===================

Sending events in a perf buffer and then having an additional map to
rate limit events is a bit silly, we could just use a map as an "events"
channel and then read from userspace.

Benefits:
- No need for two maps, lower memory usage;
- Easier code;
- Stop relying on `.Poll()` which allocates a lot of memory;

Issues:
- BPF maps just use a hash, if there are any collisions we can't store
  both keys, this might result in some events being lost. We should
calculate the probability of this happening;
- We will have to do very frequent BPF map reads. We might want to use
  batched ops.

For these reasons we will consider other alternatives in the future, but
in the meantime let's see if this fixes what some users are seeing.

Test Plan
=========

https://pprof.me/46d50b5

+ tested the flags locally

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>