bpf: Reduce iteration size by 1
Our e2e tests in CI are failing with `argument list too long`. They seem
to be running on `6.2.0-1011-azure`. So this failure is not expected but
let's see if we can address this.

Reducing the size has a small performance penalty so it would be
preferred to use a well-known kernel that works.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>