bpf: Fix task identifier usage
The name of PID (process ID) and TID (thread ID) in the kernel vs. in
userspace is different, which is a bit confusing [0].

We tried to be careful here, but in some cases we used the "per thread
ID" (using this name for clarity's sake) when we should have used the
"per process ID". While I don't believe this caused any correctness
issues, it will add overhead for multi threaded applications. For
example, in `process_info` we keep a mapping of "per process ID" to the
process information, which among other things, it contains the
executable mappings of the process, so we can find the current object we
are executing. Having this "per thread ID" is redundant, as every sibling
thread has the same memory mappings as they share the same address
space.

With these changes we will reduce CPU usage as we have to do less work
and we can keep more info in the BPF maps, as we will use them more
efficiently.

Running https://github.com/parca-dev/testdata/pull/21:
**Before**

```
$ ppstree -p `pidof basic-cpp-multithreaded-no-fp`
basic-cpp-multi(146858)─┬─{basic-cpp-multi}(146859)
                        └─{basic-cpp-multi}(146860)
```

```
$ sudo bpftool map dump name process_info | grep "key"
        "key": 146858,
        "key": 146859,
        "key": 146860,
```

**After**

```
$ sudo bpftool map dump name process_info | grep "key"
        "key": 146858,
```

Test Plan
=========

Before: https://pprof.me/2de82f2
After: https://pprof.me/4b7b713

+ integration tests + visual inspection of various profiles.

[0]: https://man7.org/linux/man-pages/man2/getpid.2.html (notes section)

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>