Remove redundant unwind rows
As we binary search over the unwind information, any neighbouring rows
with the same unwind info can be summarised in just one.

This optimisation was obvious from early on and I added it back in September
last year (parca-dev/parca-agent#794) but it was removed later on and never re-added.

Test Plan
=========

Snapshot tests + added unit tests

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>