*: Use optimized symtabs for jitdump and perfmaps
This patch utilizes the optimized symbol table previously only used for
ksyms to move the results of parsing perfmaps and jitdumps to external
memory and memory map it.

Since the overwhelming majority of symbols are never used this allows
saving a large amount of memory while still keeping those that are used
often in cache, which is handled by the operating system using mmap.