pkg/perf: Prevent intermediate perfmap/jitdump entry lists
Previously we maintained multiple lists of perfmap and jitdump entries
as well as all related strings. This caused somewhat large memory spikes
when a lot of these have to be loaded frequently.

Now strings are never on heap and are directly written to the optimized
symtab file, and only entries are loaded into memory once so they can be
sorted and deduplicated easily and then without any further intermediate
lists directly written to the optimized symtab file.

If this is also not sufficient the next optimization would be to read
perfmap files from the back and directly write entries that are read if
they don't conflict with any already written memory ranges.