pkg/objectfile: Use mount namespace id in cache key instead of build id
In the worst case, determining the build id can be very expensive, as we
need to hash the .text section. Determining the mount namespace id is
significantly cheaper and with mount namespace id, modtime and path we
still have a key that is highly stable for the properties that we need
it for.