unwinder/native: Increase size of mappings
This commit increases the maximum mappings a process without FP can
have. This limitation affected a small number of users, but it was still
a blocker for some.

Additionally, rather than a linear search, we switch to using binary
search on the mappings as otherwise it's too many instructions for the
BPF verifier to check.

Test Plan
=========

```
=== RUN   TestCPUProfilerWorks/mixed_mode_unwinding
=== RUN   TestCPUProfilerWorks/unwinder_metrics_work
--- PASS: TestCPUProfilerWorks (20.79s)
    --- PASS: TestCPUProfilerWorks/dwarf_unwinding (0.08s)
    --- PASS: TestCPUProfilerWorks/fp_unwinding (0.67s)
    --- PASS: TestCPUProfilerWorks/mixed_mode_unwinding (0.04s)
    --- PASS: TestCPUProfilerWorks/unwinder_metrics_work (0.00s)
PASS
ok      github.com/parca-dev/parca-agent/test/integration       20.810s
```

and kernel tests pass