unwinder/native: Per code region unwinder selection  (#2419)
unwinder/native: Per code region unwinder selection

Before this commit we either used frame pointers or DWARF-derived unwind
information to unwind a process. We first attempt to unwind using frame
pointers and if it looks like it worked (we can't be sure about this,
see comments for more details) we aggregate the sample, otherwise we ask
userspace to write unwind information, if present. We have to exceptions
for this case, we never generated unwind info for v8-based apps [0] nor
for Go-based apps [1].

This commit enables per code region switching of unwinders, what this
means in practice is that it will allow us to unwind Go or V8
applications, that should have frame pointers in the majority of their
code using the FP unwinder, but switch to the DWARF-derived unwinder for
part of the code that do not have them, such as for native libraries
written in C or C++ that were compiled without them. Our agent is a good
example of this as we rely on CGO with a bunch of libraries that can be
compiled without FPs.

Additionally, this work also adds support for unwinding interpreters
whose native code sections contain FPs.

- [0]: https://www.polarsignals.com/blog/posts/2023/07/12/correctly-profiling-node-js-with-zero-instrumentation
- [1]: Go has FP enabled by default

Test Plan
=========

Tested on the Agent, CI is clean,

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>