support aarch64 binaries that use PAC
Pointer authentication (PAC) is an aarch64 security feature intended
to make stack smashing/ROP attacks more difficult. For a good
introduction, see (Raymond Chen's blog post)[1].

The basic idea is that functions sign their `lr` register in the
function prologue and verify the signature in the epilogue before
returning. This signature is stored in the unused portion of the
pointer; the rest of the register is left unchanged. Instructions of type
DW_CFA_AARCH64_negate_ra_state are emitted in the .eh_frame entries to
indicate regions where this has happened.

Before the present commit, this caused two problems for parca-agent:

1. This instruction shares its opcode with DW_CFA_GNU_window_save,
   which takes a SLEB128 argument, whereas DW_CFA_AARCH64_negate_ra_state
   does not. During unwind table parsing, parca-agent would misinterpret
   DW_CFA_AARCH64_negate_ra_state as DW_CFA_GNU_window_save, attempt to
   read the SLEB128 value, end up pointing at the wrong address,
   and then generally go off the rails. This can eventually cause
   parca-agent to panic.
2. Even if we correctly interpret the unwind tables, the unwinder
   would not know what to do with an `lr` register that has had this
   transformation applied to it.

This commit fixes the first problem by simply ignoring
DW_CFA_AARCH64_negate_ra_state entries, and fixes the second problem
by canonicalizing the value of the `lr` register (or its historical
values stored on the stack) before attempting to use it.

[1]: https://devblogs.microsoft.com/oldnewthing/20220819-00/?p=107020