fix infinite loop in setUnwindTableForMapping
If a function end marker is not found within the remaining possible
size of an unwind table shard, we allocate a new shard and try
again. This leads to an infinite loop in the case that we can't find
any function end marker even within the maximum possible shard
size. This commit makes us return an error in that case instead.