Makefile fixes for guix
This makes the various `make` targets requiring `sudo` work better
for Guix-derived development environments, because it ensures that
the root shell in which they're executed is using the correct versions
of all the development tools, and has the correct environment
variables set.

I've been using this patch locally for several weeks. Normally I
wouldn't try to impose something that only benefits me on everyone else,
but since this should have no effect on "typical" dev environments
that just use the system git/gcc/etc., I hope people will be okay with it.