Surface unwind failure reasons in status page
Each time we return from the eBPF program without a stack, bump a
counter describing the reason for doing so. In user space, every time
through the profiling loop, slurp these counters and surface them in
the status page.

My hope is that this will be a first step towards helping us debug why
some processes are not being unwound in some user environments.