Fix unwinder loading on 6.8 kernels
See comments, and mailing list thread linked therein, for
details. Before this change, recent kernels (6.8 and later) cannot
verify the native unwinder, because improvements in the precision of
register bounds tracking cause an explosion in state space resulting
in exceeding the limit of 1 million instructions processed per load.

This commit works around that issue, as the comment in the code explains.