Add "unwind failed" synthetic stacks
For each unwind failed reason, add a fake stack with two frames: one
for a fake function named "unwind failed" and the other for a fake
function whose name describes the failure reason.

This will let people see the true CPU usage of their processes in the UI,
even when unwinding failed for many stacks. It will also give people an idea
that something is going wrong without having to look at the status page.