Fix off-by-one error in chunk splitting
If the PC of the "end-of-function" marker was exactly one greater than
the PC of the last real table entry, and if the unwind entries of the
function (including the end-of-function marker) were too big to fit in
the current chunk also by exactly one, then we would erroneously take
the entire function _except_ the marker, and start the next chunk with
the marker. This commit fixes that defect.