Use both .eh_frame and .debug_frame
.debug_frame is quite similar to .eh_frame, but some binaries have one
and not the other, or have a mix of things in both. In this commit, we
construct things from both.

We don't have much experience with our .debug_frame parser in the real
world, so for now if anything is correpted or fails parsing that one,
we just fall back to the old behavior of using only .eh_frame and log
a warning / bump a metric. After we gain some more confidence, we can
reconsider that logic.