Dynamically discover offsets for finding custom labels map
This reads the DWARF info of a Go program to discover the offsets of
the map containing custom pprof labels, and passes them to the
unwinder, rather than hardcoding them manually. As before, the unwinder walks
this label map to discover the OTEL thread ID, if it has been saved.

Future work will include extending this to Rust as well as making it
work in more exotic situations (e.g. when the Go runtime has been
loaded from a shared library rather than being part of the main executable)