bpf/unwinders/pyperf: Use an LRU map and increase the size
In high PID churn environments we can quickly exceed 4096 PIDs, which
currently is a hard limit for the agent. This both changes the map type
to be an LRU so newer processes get the same chance of being profiled as
old processes as well as increase the number of active python processes
to 16384.

Each map entry is 26 bytes, so the total map size is around 450kb after
this change.