Fix various issues with debuginfo uploading (#2895)
* Don't compute new build ID for extracted debuginfo

For files without a build ID, we are computing it based on a hash of
the .text section. When we extract the debug-only object file for
uploading, we can't use it to re-compute the build ID, because the
.text section is empty. We need to re-use the same build ID from the
original object file.

* Increase size of upload job queue.

If the upload job queue is full, unwind info can't be
extracted. Having this set to 128 seems to apply backpressure too
conservatively in practice, blocking unwinding for programs with a lot
of mappings.

* address issues

* [pre-commit.ci lite] apply automatic fixes

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>