Support TID (#2970)
Bump to our fork of opentelemetry-ebpf-profiler that supports thread IDs, and add the thread ID as a label.

Previously we were just getting the `comm` of the first thread we saw a sample for and saving that as the entire process's `comm`. This PR saves that as `thread_name` instead and gets the main thread's comm (i.e. process name) from procfs, saving that as `comm` instead.