Fixed file writing/reading
Previously using the Arrow IPC File format it contains hardcoded
offsets into the file. This becomes problematic when you try to write
multiple of these to a single file, and the IPC reader can't handle a
section reader (as many of the offsets will be out of bounds of the
section). Switching to using the IPC Stream format fixes that and allows
us to concatenate multiple arrow records to a single physical file and
read them back.