Allow disabling tracepoints
Currently we fail to run if tracepoints can't be set on various
sub-events of /sys/kernel/debug/tracing/events/syscalls/ . Currently
these are only used to detect process exit, so with
--disable-tracepoints we will not detect that as quickly/precisely,
leading to bloat in the BPF maps. Without tracepoints, we will only
detect exited processes every pidCleanupInterval (by default 5
minutes), through a scan of /proc. Thus setting this flag is not
recommended unless necessary.

It might be necessary because certain obscure distributions apparently
run with a configuration that doesn't expost these syscall tracepoint
events.

I have confirmed that with this patch (and setting the flag), the
agent starts up and seems to work normally on a kernel without
CONFIG_FTRACE_SYSCALLS, which fails without this patch.