reporter: Improve stack hashing for better cachability
Previously we used the internal hashes of traces from within the
opentelemetry-ebpf-profiler, however, that hash is the same whether we
end up getting full symbols for eg. interpreted frames, which meant that
we couldn't safely cache incomplete stacks.

This change introduces a separate hashing mechanism, that hashes the
fully symbolized stacks, including whether a frame was symbolized or
not, meaning this hash can safely be used to cache those stacks.

A relatively unscientific experiment with a test cluster shows roughly a
40-50% reduction in bytes sent over the wire for servers that perform
caching.

Judging by profiling data (or lack thereof), this change does not appear
to have any effect on the baseline CPU usage of the agent.